@@ -20,8 +20,9 @@
     "node-sass": "^4.5.2",
     "normalize.css": "^6.0.0",
     "postcss-modules-values": "^1.2.2",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "prop-types": "^15.5.10",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
     "react-select": "^1.0.0-rc.3",
     "sass-loader": "^4.1.1",
     "socket.io-client": "^1.7.3",
