@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 import Griddle from 'griddle-react';
 
@@ -41,8 +42,8 @@ export default class SmartGrid extends React.Component {
 }
 
 SmartGrid.propTypes = {
-    uuid: React.PropTypes.string.isRequired,
-    socket: React.PropTypes.object.isRequired,
-    columns: React.PropTypes.array.isRequired,
-    resultsPerPage: React.PropTypes.number.isRequired
+    uuid: PropTypes.string.isRequired,
+    socket: PropTypes.object.isRequired,
+    columns: PropTypes.array.isRequired,
+    resultsPerPage: PropTypes.number.isRequired
 };
