@@ -1,3 +1,4 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 import noUiSlider from 'nouislider/distribute/nouislider.min.js';
 import 'nouislider/distribute/nouislider.css';
@@ -74,26 +75,26 @@ export default class Nouislider extends React.Component {
 
 Nouislider.propTypes = {
     // http://refreshless.com/nouislider/slider-options/#section-animate
-    uuid: React.PropTypes.string.isRequired,
-    socket: React.PropTypes.object.isRequired,
-    animate: React.PropTypes.bool,
+    uuid: PropTypes.string.isRequired,
+    socket: PropTypes.object.isRequired,
+    animate: PropTypes.bool,
     // http://refreshless.com/nouislider/behaviour-option/
-    behaviour: React.PropTypes.string,
+    behaviour: PropTypes.string,
     // http://refreshless.com/nouislider/slider-options/#section-Connect
-    connect: React.PropTypes.oneOfType([
-        React.PropTypes.oneOf(['lower', 'upper']),
-        React.PropTypes.bool
+    connect: PropTypes.oneOfType([
+        PropTypes.oneOf(['lower', 'upper']),
+        PropTypes.bool
     ]),
     // http://refreshless.com/nouislider/slider-options/#section-cssPrefix
-    cssPrefix: React.PropTypes.string,
+    cssPrefix: PropTypes.string,
     // http://refreshless.com/nouislider/slider-options/#section-orientation
-    direction: React.PropTypes.oneOf(['ltr', 'rtl']),
+    direction: PropTypes.oneOf(['ltr', 'rtl']),
     // http://refreshless.com/nouislider/more/#section-disable
-    disabled: React.PropTypes.bool,
+    disabled: PropTypes.bool,
     // http://refreshless.com/nouislider/slider-options/#section-limit
-    limit: React.PropTypes.number,
+    limit: PropTypes.number,
     // http://refreshless.com/nouislider/slider-options/#section-margin
-    margin: React.PropTypes.number,
+    margin: PropTypes.number,
     // http://refreshless.com/nouislider/events-callbacks/#section-change
     // onChange: React.PropTypes.func,
     // http://refreshless.com/nouislider/events-callbacks/#section-update
@@ -101,21 +102,21 @@ Nouislider.propTypes = {
     // http://refreshless.com/nouislider/events-callbacks/#section-slide
     // onUpdate: React.PropTypes.func,
     // http://refreshless.com/nouislider/slider-options/#section-orientation
-    orientation: React.PropTypes.oneOf(['horizontal', 'vertical']),
+    orientation: PropTypes.oneOf(['horizontal', 'vertical']),
     // http://refreshless.com/nouislider/pips/
-    pips: React.PropTypes.object,
+    pips: PropTypes.object,
     // http://refreshless.com/nouislider/slider-values/#section-range
-    range: React.PropTypes.object.isRequired,
+    range: PropTypes.object.isRequired,
     // http://refreshless.com/nouislider/slider-options/#section-start
-    start: React.PropTypes.arrayOf(React.PropTypes.number).isRequired,
+    start: PropTypes.arrayOf(PropTypes.number).isRequired,
     // http://refreshless.com/nouislider/slider-options/#section-step
-    step: React.PropTypes.number,
+    step: PropTypes.number,
     // http://refreshless.com/nouislider/slider-options/#section-tooltips
-    tooltips: React.PropTypes.oneOfType([
-        React.PropTypes.bool,
-        React.PropTypes.arrayOf(
-            React.PropTypes.shape({
-                to: React.PropTypes.func
+    tooltips: PropTypes.oneOfType([
+        PropTypes.bool,
+        PropTypes.arrayOf(
+            PropTypes.shape({
+                to: PropTypes.func
             })
         )
     ])
